wings_draw_setup: Refactor pointer API
As a preparation for supporting Vertex Buffer Objects in a future
commit, we want to hide more details of #vab{} record. The first
step is replace all the functions that enable or disable pointers
by two functions:

   enable_pointers(#vab{}, [Extra]).
   disable_pointers(#vab{}, [Extra]).

By using one call to enable all pointers, we can be sure that buffers
are set up in the correct order (the VBO must be bound before any
pointers start to reference it). Since the calls are passed the
entire #vab{} record, we don't have to shoehorn the VBO identifier
into #vab.face_vs, but can put it into its own field in the record.