Teach wings_dl to understand drawing funs
Display lists are deprecated in OpenGL 3.x and later (and not supported
at all in OpenGL 3.x on the Mac). Thus, in the long term we will need to
stop using display lists.

In the short term, displays don't work well in combination with
Vertex Buffer Objects with some OpenGL drivers.

Therefore, we'll teach wings_dl to call drawing funs as a replacement
for display lists.

Note that we need not worry about garbage-collecting those funs.
Even if they happen to contain references to #vab{} records with embedded
references to VBOs, wings_dl already support garbage-collecting #vab{}
records, and wings_draw is careful to always clear display lists at the
same time that #vab{} records are cleaned.